+++ cli/code_builder/lib/main.dart
@@ -17,20 +17,23 @@ void main(List<String> args) {
         exit(1);
       }
 
-      final String componentName = args.name != null ? args.name! : UserInput.prompt(message: "Write component name");
-      final String output = args.output != null ? args.output! : UserInput.prompt(message: "Write output directory");
+      final String componentName = args.name != null ? args.name! : UserInput.prompt(message: "Component name");
+      final String output = args.output != null
+          ? args.output!
+          : UserInput.prompt(message: "Output in root directory (${args.rootDirectoryPath})");
 
       if (args.type == 'component') {
         final brickPath = Brick.path(path.join(projectRoot, 'bricks', 'vader_component'));
         final generator = await MasonGenerator.fromBrick(brickPath);
-        final target = DirectoryGeneratorTarget(Directory(path.joinAll([projectRoot, ...path.split(output)])));
+        final target =
+            DirectoryGeneratorTarget(Directory(path.joinAll([args.rootDirectoryPath, ...path.split(output)])));
         await generator.generate(target, vars: <String, dynamic>{'name': componentName, 'package': args.package});
 
         stdout.writeln("Component with name '$componentName' was successfully created.");
       } else if (args.type == 'example') {
         final brickPath = Brick.path(path.join(projectRoot, 'bricks', 'example'));
         final generator = await MasonGenerator.fromBrick(brickPath);
-        final target = DirectoryGeneratorTarget(Directory(path.joinAll([projectRoot, ...path.split(output)])));
+        final target = DirectoryGeneratorTarget(Directory(path.joinAll([args.rootDirectoryPath, ...path.split(output)])));
         await generator.generate(target, vars: <String, dynamic>{'name': componentName});
 
         stdout.writeln("Example was successfully created.");