+++ cli/code_builder/CHANGELOG.md
@@ -1,3 +1,9 @@
-## 1.0.0
+## 0.1.1
+
+- Add root directory prefix for output
+  
+
+## 0.1.0
 
 - Initial version.
+- Add component build